backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]
sqlalchemy==2.0.23
aiosqlite==0.19.0
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.12
langchain-core==0.2.38
qdrant-client>=1.11.0,<2.0.0
openai>=1.40.0,<2.0.0
pypdf==3.17.4
python-pptx==0.6.23
pillow==10.1.0
pytesseract==0.3.10
python-dotenv==1.0.0
unstructured[ppt]>=0.11.0

